+++ projects/ui-lib/content/category-progress/category-progress.component.ts
@@ -1,11 +1,11 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
-import { HighlightDirective } from '@lluc_llull/ui-lib/directives';
+import { HighlightDirective, ScrollRevealDirective } from '@lluc_llull/ui-lib/directives';
 
 @Component({
     selector: 'lib-category-progress',
     standalone: true,
-    imports: [CommonModule, HighlightDirective],
+    imports: [CommonModule, HighlightDirective, ScrollRevealDirective],
     templateUrl: './category-progress.component.html',
     styleUrl: './category-progress.component.scss',
     changeDetection: ChangeDetectionStrategy.OnPush,