+++ projects/ui-lib/content/section-intro/section-intro.component.ts
@@ -1,13 +1,13 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
-import { LinkTypeDirective } from '@lluc_llull/ui-lib/directives';
+import { LinkTypeDirective, ScrollRevealDirective } from '@lluc_llull/ui-lib/directives';
 import { UiLibButtonI } from '@lluc_llull/ui-lib/interfaces';
 import { UiIconComponent } from '@lluc_llull/ui-lib/shared';
 
 @Component({
     selector: 'lib-section-intro',
     standalone: true,
-    imports: [CommonModule, LinkTypeDirective, UiIconComponent],
+    imports: [CommonModule, LinkTypeDirective, UiIconComponent, ScrollRevealDirective],
     templateUrl: './section-intro.component.html',
     styleUrl: './section-intro.component.scss',
     changeDetection: ChangeDetectionStrategy.OnPush,