+++ projects/ui-lib/directives/index.ts
@@ -1,2 +1,3 @@
 export * from './link-type.directive';
 export * from './highlight.directive';
+export * from './scroll-reveal.directive';